Drop use of `typing.TypeAlias` in type hints (fix #310).
`typing.TypeAlias` was added in Python 3.10. In Python <= 3.9, using
`typing.TypeAlias` as a type hint confuses mypy.

To avoid this, either `typing_extensions.TypeAlias` must be used, or the
`TypeAlias` type hint may be dropped entirely in favor of relying on
type inference. Based on the current, limited usage of `TypeAlias`,
the latter is preferable.

Run mypy as part of automated tests for every supported Python version
to catch such issues in the future.